Add comprehensive unit tests for socket functionality and validation script for test coverage
- Implemented a new test suite for socket operations in `test_socket.hpp`, covering initialization, creation, binding, non-blocking mode, connection timeouts, performance, and error handling.
- Added a PowerShell script `validate_test_coverage.ps1` to validate test file existence, analyze test file sizes, compile tests, and generate a detailed report on test coverage and features.
- Enhanced test coverage with over 400 test cases, ensuring robust validation of the Atom Web module's socket functionalities and overall integration.

+++ atom/utils/text/string.cpp
@@ -475,7 +475,7 @@ auto wstringToString(std::wstring_view wstr) -> std::string {
                 }
             }
             // Handle 3-byte UTF-8 characters (2048-65535)
-            else if (wc < 0x10000) {
+            else if (static_cast<uint32_t>(wc) < 0x10000) {
                 result.push_back(static_cast<char>(0xE0 | (wc >> 12)));
                 result.push_back(static_cast<char>(0x80 | ((wc >> 6) & 0x3F)));
                 result.push_back(static_cast<char>(0x80 | (wc & 0x3F)));
